+++ libraries/mediaupload/impl/src/main/kotlin/io/element/android/libraries/mediaupload/impl/ThumbnailFactory.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.libraries.mediaupload
+package io.element.android.libraries.mediaupload.impl
 
 import android.annotation.SuppressLint
 import android.content.Context
